+++ db/threadpool/threadpool.h
@@ -40,7 +40,7 @@ namespace thread
     {
         public:
             unstarted_thread(
-                size_t s,
+                uint64_t s,
                 void (*f)(graph *, void *),
                 graph *g,
                 void *a);
@@ -49,7 +49,7 @@ namespace thread
             bool operator>(const unstarted_thread &t) const;
 
         public:
-            size_t start_time;
+            uint64_t start_time;
             void (*func)(graph *, void *);
             graph *G;
             void *arg;
@@ -57,7 +57,7 @@ namespace thread
 
     inline
     unstarted_thread :: unstarted_thread( 
-            size_t s,
+            uint64_t s,
             void (*f)(graph *, void *),
             graph *g,
             void *a)
